+++ example/.eslintrc.js
@@ -2,6 +2,15 @@ module.exports = {
   root: true,
   extends: ['expo', 'prettier'],
   plugins: ['prettier'],
+  overrides: [
+    {
+      files: ['**/*.ts', '**/*.tsx'],
+      parserOptions: {
+        project: './tsconfig.json',
+        tsconfigRootDir: __dirname,
+      },
+    },
+  ],
   env: {
     node: true,
   },
@@ -13,6 +22,7 @@ module.exports = {
     },
   },
   rules: {
+    '@typescript-eslint/no-floating-promises': 'error',
     'prettier/prettier': 'error',
   },
 };